Extract injector types out of provider
